1.0 HomeScreen and welcomeScreen completed

+++ app/src/main/java/com/golden_minute/nasim/presentation/onboarding/WelcomeScreenEvents.kt
@@ -3,5 +3,5 @@ package com.golden_minute.nasim.presentation.onboarding
 sealed class WelcomeScreenEvents {
     data class OnSearchValueChanges(val searchValue: String) : WelcomeScreenEvents()
     data object SaveData : WelcomeScreenEvents()
-    data class OnSelectItem(val lat: Float,val lon:Float) : WelcomeScreenEvents()
+    data class OnSelectItem(val lat: Double, val lon: Double) : WelcomeScreenEvents()
 }
